ai_audio_tour_agent/pyproject.toml
[project]
name = "ai-audio-tour-agent"
version = "0.1.0"
description = "AI-powered audio tour guide using Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "phidata>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "googlesearch-python>=1.2.3",
    "pycountry>=24.6.1",
    "gTTS>=2.5.0",
]
